requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
PyJWT==2.8.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
asyncpg==0.29.0
prometheus-client==0.20.0
alembic==1.13.2
langgraph==0.2.34